chore: improve Link exports
